+++ frontend/src/pages/crawls.ts
@@ -5,8 +5,8 @@ import { customElement, property, state } from "lit/decorators.js";
 import { when } from "lit/directives/when.js";
 import queryString from "query-string";
 
-import needLogin from "@/classes/decorators/needLogin";
 import type { PageChangeEvent } from "@/components/ui/pagination";
+import needLogin from "@/decorators/needLogin";
 import { CrawlStatus } from "@/features/archived-items/crawl-status";
 import type { APIPaginatedList, APIPaginationQuery } from "@/types/api";
 import type { Crawl, CrawlState } from "@/types/crawler";
